FU-54 refactor: 패키지명 및 구조 변경
handler를 config에 통합시키고, entity 대신 model이라는 패키지명을 사용함
