+++ src/main/java/com/foru/freebe/auth/service/KakaoUserRegistrationService.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
 
-import com.foru.freebe.auth.entity.KakaoUser;
+import com.foru.freebe.auth.model.KakaoUser;
 import com.foru.freebe.member.entity.Member;
 import com.foru.freebe.member.repository.MemberRepository;
 import com.foru.freebe.member.service.MemberService;
