+++ src/main/java/com/foru/freebe/config/SecurityConfig.java
@@ -8,6 +8,8 @@ import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.web.SecurityFilterChain;
 
+import com.foru.freebe.auth.config.CustomAuthenticationSuccessHandler;
+
 import lombok.RequiredArgsConstructor;
 
 @Configuration
